pijector: position independent virus code.
Adding the `pijector` work in progress code. This is a snapshot of
a continuation of `minijector` that uses the delta offset trick to make
the virus code position independent.

To make seeing the delta offset in action easier the entrypoint of
infected programs is changed to the virus code.

Note: This is **not** a complete working virus! The first call to
a kernel32.dll function in generation 1 will cause the infected binary
to segfault because the API addresses are not being correctly
calculated.