CI: check links with Lychee.
This commit adds a new Github workflow, lychee-action[0], that runs
Lychee[1] to check the rendered Hugo site has no dead links. This is
a blocking check performed for all PRs and branch pushes (with a 1d link
cache to prevent hammering sites for many updates in a day).

A separate cron based workflow that creates issues for findings will be
added in a follow-up commit to prevent rot between PRs.

[0]: https://github.com/lycheeverse/lychee-action
[1]: https://github.com/lycheeverse/lychee